thingspeak_adaptor/requirements.txt
paho-mqtt==1.6.1
requests==2.26.0
cherrypy==18.6.1
pyjwt==2.9.0